+++ src/main/java/com/trading212/code212/repositories/ProblemRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface ProblemRepository {
 
     ProblemEntity createProblem(String title, String description, String inputUrl, String outputUrl);
-    Optional<ProblemEntity> getProblemById(Long id);
+    Optional<ProblemEntity> getProblemById(int id);
     void deleteProblemById(Long id);
     List<ProblemEntity> getAllProblems();
